requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0

# YOLO / Object detection
ultralytics>=8.0.0

# Robot arm control (Dobot)
pydobot>=1.3.1

# AprilTag detection
pupil-apriltags>=1.0.4

# Calibration and transforms
scipy>=1.7.0

# Utilities
PyYAML>=6.0
loguru>=0.6.0
tqdm>=4.62.0

# Visualization
matplotlib>=3.4.0
